Add final regression test
